+++ src/index.html
@@ -8,6 +8,7 @@
       crossorigin="anonymous"
     ></script>
     <script src="./todo.js"></script>
+    <script src="./status.js"></script>
     <title>To Do list</title>
   </head>
   <body>
