+++ src/script.js
@@ -1,49 +1,3 @@
-// main.js
-
-import { markAsCompleted, markAsIncomplete, filterCompleted } from './status.js';
-
-const items = [
-  { name: 'Task 1', completed: false },
-  { name: 'Task 2', completed: false },
-  { name: 'Task 3', completed: true },
-];
-
-function getItemIndexFromCheckbox(checkbox) {
-  const itemId = checkbox.dataset.itemId;
-  for (let i = 0; i < items.length; i++) {
-    if (items[i].name === itemId) {
-      return i;
-    }
-  }
-  return -1; // item not found
-}
-
-const checkboxes = document.querySelectorAll('.completed-checkbox');
-checkboxes.forEach((checkbox) => {
-  checkbox.addEventListener('change', (event) => {
-    const itemIndex = getItemIndexFromCheckbox(event.target);
-    if (itemIndex >= 0) {
-      if (event.target.checked) {
-        markAsCompleted(items[itemIndex]);
-      } else {
-        markAsIncomplete(items[itemIndex]);
-      }
-      // update the UI to reflect the change
-    }
-  });
-});
-
-const clearCompletedButton = document.querySelector('#clear-completed-button');
-clearCompletedButton.addEventListener('click', () => {
-  items = filterCompleted(items);
-  // update the UI to reflect the change
-});
-
-localStorage.setItem('items', JSON.stringify(items));
-
-
-
-
 // Define the task array with sample tasks
 const tasks = [
   {
